fix(issue-1495): Address local pre-PR review - close the pasteable-argument invariant
The redaction gate exempted absolute, `~`, `./` and `../` spellings from query
redaction but not the BARE relative one, which `is_bare_local_source` accepts
and `resolve_recorded_local_source` resolves. A source directory named `cat?x`
was cut at the `?`. It asks the resolver's own predicates now rather than a
fourth hand-written prefix list, so the gate and the resolver cannot drift.

The URL check has to lead. `is_bare_local_source` asks
`looks_like_remote_source`, which parses, and parsing builds a display through
this very function; a scp-like spelling carries no separator, so it reaches
that question and comes back forever. `parse_remote_url` settles it without
parsing further, and costs no accuracy — a URL-shaped spelling was never a
path. Reverting just that line stack-overflows, which is what its test pins.

The available-items line was the last surface composing its own `vstack add`
outside `restore_source_argument`. Its source is redacted at construction, so
the decision could not be made downstream: the offer now carries the argument
alongside the display form, decided from the raw string where the remedy's is,
and `is_pasteable_source_argument` is the one function both ask. On a healthy
source named `cat?x` the offer read `vstack add 'cat?<redacted>'` — a
paste-ready command naming a directory that does not exist, under a ✓ and an
exit 0. An offer whose source genuinely cannot be an argument still announces
its items, without a command.

Two seams split out to keep the ratchet moving down: `check::available` and
`check::render::tests::remedy_commands`.
