+++ tools/size-ratchet-baseline.tsv
@@ -5,7 +5,7 @@ cli/src/commands/add.rs	1681
 cli/src/commands/add/source_option_tests.rs	913
 cli/src/commands/add/source_option_tests/project_install_tests.rs	438
 cli/src/commands/apply.rs	2743
-cli/src/commands/check.rs	632
+cli/src/commands/check.rs	577
 cli/src/commands/check/tests.rs	877
 cli/src/commands/refresh.rs	1525
 cli/src/commands/refresh/tests.rs	1749
